+++ box/__init__.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+# -*- coding: UTF-8 -*-
+
+__author__ = 'Chris Griffith'
+__version__ = '4.0.0'
+
+from box.box import Box
+from box.box_list import BoxList
+from box.config_box import ConfigBox
+from box.shorthand_box import SBox
+from box.exceptions import BoxError, BoxKeyError
