+++ setup.py
@@ -25,7 +25,7 @@ setup(
     license='MIT',
     author=attrs['author'],
     tests_require=["pytest", "coverage >= 3.6", "pytest-cov"],
-    install_requires=[],
+    install_requires=['toml', 'ruamel.yaml'],
     author_email='chris@cdgriffith.com',
     description='Advanced Python dictionaries with dot notation access',
     long_description=long_description,
@@ -36,11 +36,7 @@ setup(
     setup_requires=['pytest-runner'],
     classifiers=[
         'Programming Language :: Python',
-        'Programming Language :: Python :: 2',
-        'Programming Language :: Python :: 2.7',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.3',
-        'Programming Language :: Python :: 3.4',
         'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
